Module artifact::logger

Structs

Logger
A logger within the Artifact logging library. This struct is somewhat similar to an address. The struct itself only stores the name of the logger. However, the initialization functions tell the backend what kind of logger it is, what level it logs at, and other interesting information.

Enums

LoggerOutput
Indicates what kind of output stream a logger will use.