Enum artifact::logger::LoggerOutput

pub enum LoggerOutput {
    FileLog(PathBuf),
    StdoutLog,
    StderrLog,
    MultiLog(Vec<String>),
}

Indicates what kind of output stream a logger will use.

Variants

FileLog

StdoutLog

StderrLog

MultiLog

Log to various other loggers. Any messages sent to this logger will be forwarded on to the loggers with the names given. Note that messages are filtered both by the level of this logger and the level of the sub loggers assigned to it.